feat: split index.js to separate files and refactor
