+++ lib/feed.js
@@ -2,33 +2,38 @@
 
 const Parser = require('rss-parser');
 
-const feed = module.exports;
 
-feed.getItems = async function (feedUrl, entriesToPull) {
-	entriesToPull = parseInt(entriesToPull, 10) || 4;
+const Feed = {};
+
+Feed.DEFAULT_ENTRIES_TO_PULL = 4;
+
+Feed.getItems = async function (feedUrl, entriesToPull = Feed.DEFAULT_ENTRIES_TO_PULL) {
+	entriesToPull = parseInt(entriesToPull, 10) || Feed.DEFAULT_ENTRIES_TO_PULL;
 	feedUrl = `${feedUrl}?t=${Date.now()}`;
 
 	const parser = new Parser();
 	const feed = await parser.parseURL(feedUrl);
 
-	feed.items = feed.items.filter(Boolean).slice(0, entriesToPull);
-	feed.items = feed.items.map(item => ({
-		title: item.title,
-		published: item.pubDate,
-		link: { href: item.link },
-		id: item.guid || item.id,
-		tags: item.categories,
-	}));
-	return feed.items;
+	return feed.items
+		.filter(Boolean)
+		.slice(0, entriesToPull)
+		.map(item => ({
+			title: item.title,
+			published: item.pubDate,
+			link: { href: item.link },
+			id: item.guid || item.id,
+			tags: item.categories,
+		}));
 };
 
-feed.checkFeed = async function (req, res) {
+Feed.checkFeed = async function (req, res) {
 	if (!req.query.url) {
 		return res.json('Please enter feed url!');
 	}
+
 	let entries;
 	try {
-		entries = await feed.getItems(req.query.url, 3);
+		entries = await Feed.getItems(req.query.url, 3);
 	} catch (err) {
 		return res.json(err.message);
 	}
@@ -57,5 +62,8 @@ feed.checkFeed = async function (req, res) {
 
 		return entry;
 	});
+
 	res.json(entries);
 };
+
+module.exports = Feed;
