+++ lib/jobs.js
@@ -7,6 +7,7 @@ const nconf = require.main.require('nconf');
 const plugins = require.main.require('./src/plugins');
 const pubsub = require.main.require('./src/pubsub');
 
+
 let pullFeedsInterval;
 
 const cronJobs = [
@@ -18,8 +19,9 @@ const cronJobs = [
 	new cron('0 0 0 0 0 6', (() => { pullFeedsInterval(60 * 24 * 7); }), null, false), // one week
 ];
 
-const jobs = module.exports;
-jobs.init = function (pullMethod) {
+const Jobs = {};
+
+Jobs.init = function (pullMethod) {
 	pullFeedsInterval = pullMethod;
 };
 
@@ -53,3 +55,5 @@ function stopCronJobs() {
 pubsub.on('nodebb-plugin-rss:deactivate', () => {
 	stopCronJobs();
 });
+
+module.exports = Jobs;
